Moved edition toggler to editor/components + applied for style forms
