+++ lib/assets/javascripts/cartodb3/editor/components/edition-toggle/edition-toggle-panel-view.js
@@ -1,9 +1,8 @@
 var cdb = require('cartodb.js');
-var _ = require('underscore');
-var TabPaneView = require('../tab-pane/tab-pane-view');
-var TabPaneCollection = require('../tab-pane/tab-pane-collection');
+var TabPaneView = require('../../../components/tab-pane/tab-pane-view');
+var TabPaneCollection = require('../../../components/tab-pane/tab-pane-collection');
 var EditionToggleBarView = require('./edition-toggle-bar-view');
-var Template = require('./edition-toggle-panel.tpl');
+var template = require('./edition-toggle-panel.tpl');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
@@ -19,18 +18,15 @@ module.exports = cdb.core.View.extend({
   },
 
   render: function () {
-    var self = this;
-    this._layerTabPaneView = new TabPaneView(
-      _.extend({
-        collection: self.collection,
-        template: function () {
-          return Template;
-        }
-      })
-    );
+    var layerTabPaneView = new TabPaneView({
+      collection: this.collection,
+      template: function () {
+        return template;
+      }
+    });
 
-    this.$el.append(this._layerTabPaneView.render().$el);
-    this.addView(this._layerTabPaneView);
+    this.$el.append(layerTabPaneView.render().el);
+    this.addView(layerTabPaneView);
 
     var editionBarView = new EditionToggleBarView({
       collection: this.collection
