+++ lib/assets/javascripts/cartodb3/editor/components/edition-toggle/edition-toggle-panel.tpl
