+++ lib/assets/javascripts/cartodb3/editor/components/toggler/toggler-view.js
@@ -15,11 +15,10 @@ module.exports = cdb.core.View.extend({
     }
 
     this.listenTo(this.collection, 'change', this.render);
-    this._labels = this.collection.map(function (m) {
+    this.add_related_model(this.collection);
+    this._labelsArray = this.collection.map(function (m) {
       return m.get('label');
     });
-
-    this.render();
   },
 
   render: function () {
@@ -27,7 +26,7 @@ module.exports = cdb.core.View.extend({
 
     this.$el.html(
       template({
-        labels: this._labels,
+        labels: this._labelsArray,
         checked: checked
       })
     );
