+++ lib/assets/javascripts/cartodb3/editor/components/toggler/toggler.tpl
