+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/infowindow/infowindow-view.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb.js');
-var EditionTogglePanelView = require('../../../../components/edition-toggle/edition-toggle-panel-view');
-var CodeMirrorView = require('../../../../components/code-mirror/code-mirror-view');
+var EditionTogglePanelView = require('../../../../editor/components/edition-toggle/edition-toggle-panel-view');
+var CodeMirrorView = require('../../../../editor/components/code-mirror/code-mirror-view');
 var InfowindowContentView = require('./infowindow-content-view');
 
 module.exports = cdb.core.View.extend({
