+++ lib/assets/javascripts/cartodb3/editor/style/style-cartocss-view.js
@@ -1,5 +1,5 @@
   var cdb = require('cartodb.js');
-var CodeMirrorView = require('../../components/code-mirror/code-mirror-view');
+var CodeMirrorView = require('../../editor/components/code-mirror/code-mirror-view');
 
 module.exports = cdb.core.View.extend({
