+++ lib/assets/javascripts/cartodb3/editor/style/style-view.js
@@ -1,5 +1,5 @@
 var cdb = require('cartodb.js');
-var EditionTogglePanelView = require('../../components/edition-toggle/edition-toggle-panel-view');
+var EditionTogglePanelView = require('../../editor/components/edition-toggle/edition-toggle-panel-view');
 var StyleContentView = require('./style-content-view');
 var StyleCartoCSSView = require('./style-cartocss-view');
 
@@ -43,6 +43,9 @@ module.exports = cdb.core.View.extend({
           modals: self._modals,
           querySchemaModel: self._querySchemaModel
         });
+      },
+      createControlView: function () {
+        return new cdb.core.View();
       }
     }, {
       label: _t('editor.style.style-toggle.cartocss'),
@@ -52,12 +55,14 @@ module.exports = cdb.core.View.extend({
           layerDefinitionModel: self._layerDefinitionModel,
           styleModel: self._styleModel
         });
+      },
+      createControlView: function () {
+        return new cdb.core.View();
       }
     }];
 
     var editionTogglePanelView = new EditionTogglePanelView({
-      panes: tabPaneTabs,
-      controlsView: new cdb.core.View()
+      panes: tabPaneTabs
     });
 
     this.$el.append(editionTogglePanelView.render().el);
