+++ lib/assets/test/spec/cartodb3/editor/components/edition-toggle/edition-toggle-bar-view.spec.js
@@ -1,10 +1,8 @@
 var $ = require('jquery');
-var Backbone = require('backbone');
-Backbone.$ = $;
-var EditionTogglerPanelView = require('../../../../../javascripts/cartodb3/components/edition-toggle/edition-toggle-bar-view.js');
-var TabPaneCollection = require('../../../../../javascripts/cartodb3/components/tab-pane/tab-pane-collection');
+var EditionTogglerPanelView = require('../../../../../../javascripts/cartodb3/editor/components/edition-toggle/edition-toggle-bar-view.js');
+var TabPaneCollection = require('../../../../../../javascripts/cartodb3/components/tab-pane/tab-pane-collection');
 
-describe('components/edition-toggle/edition-toggle-bar-view', function () {
+describe('editor/components/edition-toggle/edition-toggle-bar-view', function () {
   var Dummy = cdb.core.View.extend({
     render: function () {
       this.$el.html(this.options.content);
