+++ lib/assets/test/spec/cartodb3/editor/components/edition-toggle/edition-toggle-panel-view.spec.js
@@ -1,9 +1,7 @@
 var $ = require('jquery');
-var Backbone = require('backbone');
-Backbone.$ = $;
-var EditionTogglerPanelView = require('../../../../../javascripts/cartodb3/components/edition-toggle/edition-toggle-panel-view.js');
+var EditionTogglerPanelView = require('../../../../../../javascripts/cartodb3/editor/components/edition-toggle/edition-toggle-panel-view.js');
 
-describe('components/edition-toggle/edition-toggle-panel-view', function () {
+describe('editor/components/edition-toggle/edition-toggle-panel-view', function () {
   beforeEach(function () {
     var Dummy = cdb.core.View.extend({
       render: function () {
