+++ lib/assets/test/spec/cartodb3/editor/components/toggler/toggler-view.spec.js
@@ -1,10 +1,8 @@
 var $ = require('jquery');
-var Backbone = require('backbone');
-Backbone.$ = $;
-var Toggler = require('../../../../../javascripts/cartodb3/components/toggler/toggler-view.js');
-var TabPaneCollection = require('../../../../../javascripts/cartodb3/components/tab-pane/tab-pane-collection');
+var Toggler = require('../../../../../../javascripts/cartodb3/editor/components/toggler/toggler-view.js');
+var TabPaneCollection = require('../../../../../../javascripts/cartodb3/components/tab-pane/tab-pane-collection');
 
-describe('components/toggler/toggler', function () {
+describe('editor/components/toggler/toggler', function () {
   var Dummy = cdb.core.View.extend({
     render: function () {
       this.$el.html(this.options.content);
@@ -40,10 +38,6 @@ describe('components/toggler/toggler', function () {
     this.view.render();
   });
 
-  it('should have no leaks', function () {
-    expect(this.view).toHaveNoLeaks();
-  });
-
   it('should render properly', function () {
     expect(this.view.$('.js-input').length).toBe(1);
     expect(this.view.$('.js-input').prop('checked')).toBe(false);
@@ -53,6 +47,10 @@ describe('components/toggler/toggler', function () {
   it('should re render on change collection', function () {
     collection.at(1).set({selected: true});
     expect(this.view.render).toHaveBeenCalled();
+  });
+
+  it('should have no leaks', function () {
+    expect(this.view).toHaveNoLeaks();
   });
 
   afterEach(function () {
